% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dboztools-package.R
\docType{package}
\name{dboztools}
\alias{dboztools}
\alias{dboztools-package}
\title{My Utilities Package}
\description{
A collection of commonly-used functions and RMarkdown templates.
}
\author{
\strong{Maintainer}: Dave Bosworth \email{David.Bosworth@water.ca.gov} (\href{https://orcid.org/0000-0003-0740-3390}{ORCID})

}
\keyword{internal}
